Created Node.cpp and Node.h files
Created a outline for the Node that will be used in the neural network. Currently it takes in a vector of input values passed by reference and uses this to calculate an weight sum which is then scaled to an activation amount. This can be then used to compute output, which the neural net can send to all other nodes. The other nodes can accept all these inputs as a vector and the cycle will continue.